Begin work on Dragon and DragonTree structure
Began working on the DragonTree structure, which meant I had to go and work on the
Dragon structure, which meant I had to go back and test some stuff in DragonTree since
they are all connected.

ANYWAY I got the basic format of Dragon down, so all that remains is to add Get and Set
methods for all the different parameters within the DRG format.  Simple stuff.

On the side of the DragonTree, lots still needs to be done.  I have been bad and have
been testing in the init function.  However I know how to save files now, so we just
need to move those tested bits to their own functions and come up with the other necessary
structure and functions.  Fun.