Finish get and set methods
All get and set methods for the DRG tags are now in place.
For the list style tags (the ancestor and descendant tags), all data is deleted and then replaced
with the new data.  There may be a better and more efficient way to do it, but I cannot think of it now.
Also the gene and color methods were added and compacted into one method for all three levels.