Began User Interface
Began constructing a basic user interface.
Currently consists of a window with three panels in it.

Two panels live in the Display object, which is the view of our MVC.
Will have lots of functions to update and change what is displayed.

DAL9001 object creates the window and a side panel with buttons in it.
Buttons as of yet do nothing, but this will be the controller/main file when that day rolls around.